refactor(web): encapsulate review Lottie runtime
